requirements.txt
requests~=2.32.3
psycopg~=3.2.6
psycopg-binary~=3.2.6
apache-airflow~=2.10.5